refactor(ghostvis): split config/snapshot parsing helpers into GhostVisualBuilder.Parsing.cs
